PR comments: mv fixprime generator to random-prime.h ; use bit generator ;
